Move arithmetic_scalar -> scalar/arithmetic,relational
