+++ temporian/core/operators/BUILD
@@ -24,7 +24,6 @@ py_library(
         ":set_index",
         ":since_last",
         ":unique_timestamps",
-        "//temporian/core/operators/arithmetic_scalar",
         "//temporian/core/operators/binary",
         "//temporian/core/operators/calendar:day_of_month",
         "//temporian/core/operators/calendar:day_of_week",
@@ -35,6 +34,7 @@ py_library(
         "//temporian/core/operators/calendar:month",
         "//temporian/core/operators/calendar:second",
         "//temporian/core/operators/calendar:year",
+        "//temporian/core/operators/scalar",
         "//temporian/core/operators/window:moving_count",
         "//temporian/core/operators/window:moving_max",
         "//temporian/core/operators/window:moving_min",
