--- temporian/core/operators/arithmetic_scalar/__init__.py
@@ -1,41 +1,0 @@
-# pylint: disable=unused-import
-
-# TODO: Is this file needed?
-
-from temporian.core.operators.arithmetic_scalar.add import (
-    add_scalar,
-    AddScalarOperator,
-)
-from temporian.core.operators.arithmetic_scalar.divide import (
-    divide_scalar,
-    DivideScalarOperator,
-)
-from temporian.core.operators.arithmetic_scalar.floordiv import (
-    floordiv_scalar,
-    FloorDivScalarOperator,
-)
-from temporian.core.operators.arithmetic_scalar.subtract import (
-    subtract_scalar,
-    SubtractScalarOperator,
-)
-from temporian.core.operators.arithmetic_scalar.multiply import (
-    multiply_scalar,
-    MultiplyScalarOperator,
-)
-
-from temporian.core.operators.arithmetic_scalar.equal import (
-    equal_scalar,
-    EqualScalarOperator,
-)
-
-from temporian.core.operators.arithmetic_scalar.greater import (
-    greater_scalar,
-    GreaterScalarOperator,
-)
-
-from temporian.core.operators.arithmetic_scalar.less import (
-    less_scalar,
-    LessScalarOperator,
-)
-
-# pylint: enable=unused-import
