+++ temporian/core/operators/scalar/BUILD
@@ -1,0 +1,54 @@
+package(
+    default_visibility = ["//visibility:public"],
+    licenses = ["notice"],
+)
+
+# Libraries
+# =========
+
+py_library(
+    name = "scalar",
+    srcs = ["__init__.py"],
+    srcs_version = "PY3",
+    deps = [
+        ":arithmetic",
+        ":base",
+        ":relational",
+    ],
+)
+
+py_library(
+    name = "base",
+    srcs = ["base.py"],
+    srcs_version = "PY3",
+    deps = [
+        "//temporian/core/data:feature",
+        "//temporian/core/data:node",
+        "//temporian/core/operators:base",
+        "//temporian/proto:core_py_proto",
+    ],
+)
+
+py_library(
+    name = "arithmetic",
+    srcs = ["arithmetic.py"],
+    srcs_version = "PY3",
+    deps = [
+        ":base",
+        "//temporian/core:operator_lib",
+        "//temporian/core/data:node",
+    ],
+)
+
+py_library(
+    name = "relational",
+    srcs = ["relational.py"],
+    srcs_version = "PY3",
+    deps = [
+        ":base",
+        "//temporian/core:operator_lib",
+        "//temporian/core/data:dtype",
+        "//temporian/core/data:feature",
+        "//temporian/core/data:node",
+    ],
+)
