+++ temporian/core/operators/scalar/__init__.py
@@ -1,0 +1,21 @@
+# pylint: disable=unused-import
+
+from temporian.core.operators.scalar.arithmetic import (
+    add_scalar,
+    subtract_scalar,
+    multiply_scalar,
+    divide_scalar,
+    floordiv_scalar,
+    AddScalarOperator,
+    SubtractScalarOperator,
+    MultiplyScalarOperator,
+    DivideScalarOperator,
+    FloorDivScalarOperator,
+)
+
+from temporian.core.operators.scalar.relational import (
+    equal_scalar,
+    EqualScalarOperator,
+)
+
+# pylint: enable=unused-import
