+++ temporian/core/operators/scalar/base.py
