+++ temporian/implementation/numpy/operators/BUILD
@@ -25,15 +25,6 @@ py_library(
         ":set_index",
         ":since_last",
         ":unique_timestamps",
-        "//temporian/implementation/numpy/operators/arithmetic_scalar",
-        "//temporian/implementation/numpy/operators/arithmetic_scalar:add",
-        "//temporian/implementation/numpy/operators/arithmetic_scalar:divide",
-        "//temporian/implementation/numpy/operators/arithmetic_scalar:equal",
-        "//temporian/implementation/numpy/operators/arithmetic_scalar:floordiv",
-        "//temporian/implementation/numpy/operators/arithmetic_scalar:greater",
-        "//temporian/implementation/numpy/operators/arithmetic_scalar:less",
-        "//temporian/implementation/numpy/operators/arithmetic_scalar:multiply",
-        "//temporian/implementation/numpy/operators/arithmetic_scalar:subtract",
         "//temporian/implementation/numpy/operators/binary",
         "//temporian/implementation/numpy/operators/calendar:day_of_month",
         "//temporian/implementation/numpy/operators/calendar:day_of_week",
@@ -44,6 +35,7 @@ py_library(
         "//temporian/implementation/numpy/operators/calendar:month",
         "//temporian/implementation/numpy/operators/calendar:second",
         "//temporian/implementation/numpy/operators/calendar:year",
+        "//temporian/implementation/numpy/operators/scalar",
         "//temporian/implementation/numpy/operators/window:moving_count",
         "//temporian/implementation/numpy/operators/window:moving_max",
         "//temporian/implementation/numpy/operators/window:moving_min",
