+++ temporian/implementation/numpy/operators/scalar/__init__.py
@@ -1,0 +1,14 @@
+# pylint: disable=unused-import
+from temporian.implementation.numpy.operators.scalar.arithmetic import (
+    AddScalarNumpyImplementation,
+    SubtractScalarNumpyImplementation,
+    MultiplyScalarNumpyImplementation,
+    DivideScalarNumpyImplementation,
+    FloorDivideScalarNumpyImplementation,
+)
+
+from temporian.implementation.numpy.operators.scalar.relational import (
+    EqualScalarNumpyImplementation,
+)
+
+# pylint: enable=unused-import
